refactor: View의 책임에 따른 인터페이스 제거
View는 어디로 무엇을 입출력 할지 책임이 있다고 생각했다.
하지만 입력과 출력은 System.in/out이 담당한다.
우리는 System.in/out을 사용하여 어떤 형식의 글자를 전달할지 책임만 지면 된다. (Formatter)
그리고 그 역할은 쉽게 바뀌지 않으므로 인터페이스를 제거한다.
