+++ src/main/java/lotto/view/InputView.java
@@ -1,13 +1,32 @@
 package lotto.view;
 
+import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
+import java.util.Scanner;
+import lotto.util.Parser;
 
-public interface InputView {
-    long readPurchaseCost();
+public class InputView {
+    public long readPurchaseCost() {
+        System.out.println("구입금액을 입력해 주세요.");
+        return Parser.toLong(Console.readLine());
+    }
 
-    List<Integer> readWinningNumbers();
+    public List<Integer> readWinningNumbers() {
+        System.out.println("당첨 번호를 입력해 주세요.");
+        String values = Console.readLine();
+        Scanner scanner = new Scanner(System.in);
 
-    Integer readBonusNumber();
+        return Parser.split(values).stream()
+                .map(Parser::toInteger)
+                .toList();
+    }
 
-    void close();
+    public Integer readBonusNumber() {
+        System.out.println(System.lineSeparator() + "보너스 번호를 입력해 주세요.");
+        return Parser.toInteger(Console.readLine());
+    }
+
+    public void close() {
+        Console.close();
+    }
 }
